Remove unused bin/pipeline.rb file
Move all of the functions in here, which weren't
there already, into the lib/cp_env folder
