+++ lib/cp_env/pipeline.rb
@@ -59,6 +59,15 @@ def apply_namespace_dir(cluster, dir)
   apply_terraform(cluster, namespace, dir)
 end
 
+def apply_gitops_namespace_dir(cluster, dir, team_name)
+  return unless FileTest.directory?(dir)
+
+  namespace = File.basename(dir)
+  apply_kubernetes_files(cluster, namespace, dir)
+  apply_gitops_kubernetes_files(cluster, team_name, dir)
+  apply_terraform(cluster, namespace, dir)
+end
+
 def plan_namespace_dir(cluster, dir)
   return unless FileTest.directory?(dir)
 
@@ -73,6 +82,11 @@ def apply_kubernetes_files(_cluster, namespace, dir)
   end
 end
 
+def apply_gitops_kubernetes_files(_cluster, team_name, dir)
+  log("green", "applying concourse-#{team_name}")
+  execute("kubectl -n concourse-#{team_name} apply -f #{dir}/gitops-resources")
+end
+
 def contains_kubernetes_files?(dir)
   Dir.glob("#{dir}/*.{yaml,yml,json}").any?
 end
@@ -81,8 +95,8 @@ def apply_terraform(cluster, namespace, dir)
   Terraform.new(cluster: cluster, namespace: namespace, dir: dir).apply
 end
 
-def execute(cmd, can_fail: false)
-  log("blue", "executing: #{cmd}")
+def execute(cmd, can_fail: false, silent: false)
+  log("blue", "executing: #{cmd}") unless silent
   stdout, stderr, status = Open3.capture3(cmd)
 
   unless can_fail || status.success?
@@ -91,7 +105,7 @@ def execute(cmd, can_fail: false)
     raise
   end
 
-  puts stdout
+  puts stdout unless silent
 
   [stdout, stderr, status]
 end
